[QuyenLe] edit database, random data for db, FUNCTION_PROCEDURE, code KiemHang

+++ PawnShopManager/PawnShopManager/Entity/ChuocHang.cs
@@ -7,10 +7,10 @@ namespace PawnShopManager.Entity
       public Int32 chuoc_Id { get; set; }
       public Int32 hd_Id { get; set; }
       public String chuoc_NgayTao { get; set; }
-      public Int16 chuoc_SoNgay { get; set; }
-      public float chuoc_TienLai_ThoaThuan { get; set; }
-      public float chuoc_TienLai_QuaHan { get; set; }
-      public float chuoc_TongTien { get; set; }
-      public float chuoc_TangGiam { get; set; }
+      public Int32 chuoc_SoNgay { get; set; }
+      public Double chuoc_TienLai_ThoaThuan { get; set; }
+      public Double chuoc_TienLai_QuaHan { get; set; }
+      public Double chuoc_TongTien { get; set; }
+      public Double chuoc_TangGiam { get; set; }
    }
 }
