+++ PawnShopManager/PawnShopManager/Entity/TraLaiTruoc.cs
@@ -7,8 +7,8 @@ namespace PawnShopManager.Entity
       public Int32 tralai_Id { get; set; }
       public Int32 hd_Id { get; set; }
       public String tralai_NgayTao { get; set; }
-      public Int16 tralai_So_Ngay { get; set; }
-      public float tralai_Tien { get; set; }
+      public Int32 tralai_So_Ngay { get; set; }
+      public Double tralai_Tien { get; set; }
 
    }
 }
